Project Restructure to handle Components
Due to one of my main features which includes clicking on buttons in
respective player's text channels, the project structure should be
slightly changed :(. But, I genuinely think that this is the best way
for me to go from here. This doesn't affect most people's changes in
other slash commands.