Implement Ship Placement Feedback
The bot now tells the user whether their placement selections are valid
or not. I decided that a button or feedback message is sent after the
user has made a set of selections. If it's a valid placement a button
is sent. If it's invalid, then a simple message stating that the user
should change their selections is sent.

The actual place button is not handled tho.